Adjust the behavior of `element-to-config` codemod (#270)
* Move v4.0.0 specific utils

* Preserve `control` widget property

* Refactor `isExpressionKind`

* WIP import specifier

* Add element import specifier removal functionality

* Add element name expectations for some handlers

* Rush change
