+++ tools/codemod/src/utils/TransformUtils.ts
@@ -3,8 +3,6 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import { ObjectExpression, ObjectProperty, JSCodeshift, ASTPath, JSXElement, JSXIdentifier, Identifier, JSXMemberExpression, MemberExpression, SpreadProperty, ASTNode, RestElement, Property, Collection, ExpressionStatement, ArrayExpression, ConditionalExpression } from "jscodeshift";
-import { ObjectExpressionCollection, useObjectExpression } from "./ObjectExpression";
-import { transformAbstractWidget } from "../v4.0.0/widget";
 
 export type ExpressionKind = ExpressionStatement["expression"];
 export type PatternKind = RestElement["argument"];
@@ -29,14 +27,8 @@ export interface UniFormat {
   spreadAttr: ExpressionKind[];
 }
 
-export const expressionKindTypes = new Set(["Identifier", "FunctionExpression", "ThisExpression", "ArrayExpression", "ObjectExpression", "Literal", "SequenceExpression", "UnaryExpression", "BinaryExpression", "AssignmentExpression", "MemberExpression", "UpdateExpression", "LogicalExpression", "ConditionalExpression", "NewExpression", "CallExpression", "ArrowFunctionExpression", "YieldExpression", "GeneratorExpression", "ComprehensionExpression", "ClassExpression", "Super", "TaggedTemplateExpression", "TemplateLiteral", "MetaProperty", "AwaitExpression", "ImportExpression", "ChainExpression", "OptionalCallExpression", "OptionalMemberExpression", "JSXIdentifier", "JSXExpressionContainer", "JSXElement", "JSXFragment", "JSXMemberExpression", "JSXText", "PrivateName", "TypeCastExpression", "DoExpression", "BindExpression", "ParenthesizedExpression", "DirectiveLiteral", "StringLiteral", "NumericLiteral", "BigIntLiteral", "NullLiteral", "BooleanLiteral", "RegExpLiteral", "Import", "TSAsExpression", "TSNonNullExpression", "TSTypeParameter", "TSTypeAssertion"]);
-export function isExpressionKind(source: ASTNode): source is ExpressionKind {
-  return expressionKindTypes.has(source.type);
-}
-
-export const patternKindTypes = new Set(["Identifier", "RestElement", "SpreadElementPattern", "PropertyPattern", "ObjectPattern", "ArrayPattern", "AssignmentPattern", "SpreadPropertyPattern", "JSXIdentifier", "PrivateName", "TSAsExpression", "TSNonNullExpression", "TSTypeParameter", "TSTypeAssertion", "TSParameterProperty"]);
-export function isPatternKind(source: ASTNode): source is PatternKind {
-  return patternKindTypes.has(source.type);
+export function isExpressionKind(j: JSCodeshift, source: ASTNode): source is ExpressionKind {
+  return j(source).isOfType(j.Expression);
 }
 
 export function isSpecifiedConfigExpression(source: ASTNode, name: string): source is ConfigExpression {
@@ -123,157 +115,6 @@ export function getObjectProperty(j: JSCodeshift, expression: ObjectExpression, 
     // TODO: log warning
   }
   return undefined;
-}
-
-export function createEmptyConfig(j: JSCodeshift, configName: JSXIdentifier): ConfigExpression {
-  return { ...j.objectExpression([]), configName };
-}
-
-export function handleToolWidget(j: JSCodeshift, expression: ExpressionKind): ExpressionKind {
-  function handleExpression(expr: ExpressionKind): ExpressionKind {
-    let newExpr: ExpressionKind | undefined = undefined;
-    if (!newExpr && expr.type === "ConditionalExpression")
-      newExpr = handleConditionalExpression(j, expr, handleExpression);
-    if (!newExpr && expr.type === "ArrayExpression")
-      newExpr = takeFirstArrayExpressionElement(j, expr);
-    if (!newExpr) {
-      // TODO: log warning
-      return expr;
-    }
-    return newExpr;
-  }
-  return handleExpression(expression);
-}
-
-export function buildDefaultStagePanelConfigExpression(j: JSCodeshift): ConfigExpression {
-  return { ...j.objectExpression([]), configName: j.jsxIdentifier("StagePanel") };
-}
-
-export function extractSectionsFromPanelZones(j: JSCodeshift, expression: ObjectExpression) {
-  const result: [ExpressionKind | undefined, ExpressionKind | undefined] = [undefined, undefined]
-
-  const start = getObjectProperty(j, expression, "start")?.value;
-  if (start && start.type === "ObjectExpression") {
-    const startWidgets = getObjectProperty(j, start, "widgets")?.value;
-    if (startWidgets && isExpressionKind(startWidgets))
-      result[0] = startWidgets;
-  }
-  else if (start && isExpressionKind(start)) {
-    result[0] = j.memberExpression(start, j.identifier("widgets"));
-  }
-
-  const middle = getObjectProperty(j, expression, "middle")?.value;
-  if (middle && middle.type === "ObjectExpression") {
-    const middleWidgets = getObjectProperty(j, middle, "widgets")?.value;
-    if (middleWidgets && isExpressionKind(middleWidgets)) {
-      if (result[0])
-        result[0] = concatExpressions(j, result[0], middleWidgets);
-      else
-        result[0] = middleWidgets;
-    }
-  }
-  else if (middle && isExpressionKind(middle)) {
-    const middleMemberExpr = j.memberExpression(middle, j.identifier("widgets"))
-    if (result[0])
-      result[0] = concatExpressions(j, result[0], middleMemberExpr);
-    else
-      result[0] = middleMemberExpr;
-  }
-
-  const end = getObjectProperty(j, expression, "end")?.value;
-  if (end && end.type === "ObjectExpression") {
-    const endWidgets = getObjectProperty(j, end, "widgets")?.value;
-    if (endWidgets && isExpressionKind(endWidgets))
-      result[1] = endWidgets;
-  }
-  else if (end && isExpressionKind(end)) {
-    result[1] = j.memberExpression(end, j.identifier("widgets"));
-  }
-
-  return result;
-}
-
-export function appendStagePanelSection(j: JSCodeshift, stagePanel: ConfigExpression, sectionToAppend: "start" | "end", exprToAppend: ExpressionKind): "Success" | undefined {
-  let sections = getObjectProperty(j, stagePanel, "sections");
-  if (!sections) {
-    sections = j.objectProperty(j.identifier("sections"), j.objectExpression([]));
-    stagePanel.properties.push(sections);
-  }
-  if (sections.value.type !== "ObjectExpression") {
-    // TODO: log warning
-    return undefined;
-  }
-
-  let section = getObjectProperty(j, sections.value, sectionToAppend);
-  if (!section) {
-    section = j.objectProperty(j.identifier(sectionToAppend), j.arrayExpression([]));
-    sections.value.properties.push(section);
-  }
-  if (!isExpressionKind(section.value)) {
-    // TODO: log warning
-    return undefined;
-  }
-
-  section.value = concatExpressions(j, section.value, exprToAppend);
-  return "Success";
-}
-
-export function getStagePanelSectionProperty(j: JSCodeshift, stagePanel: ConfigExpression, sectionToGet: "start" | "end"): ObjectProperty | Property | undefined {
-  const sections = getObjectProperty(j, stagePanel, "sections");
-  if (!sections)
-    return undefined;
-  if (sections.value.type !== "ObjectExpression") {
-    // TODO: log warning
-    return undefined;
-  }
-
-  const section = getObjectProperty(j, sections.value, sectionToGet);
-  if (!section)
-    return undefined;
-  if (!isExpressionKind(section.value)) {
-    // TODO: log warning
-    return undefined;
-  }
-
-  return section;
-}
-
-export function replaceWidgetObjectExpression(j: JSCodeshift, widget: Collection<ObjectExpression>) {
-  useObjectExpression(j);
-  transformAbstractWidget(j, widget as ObjectExpressionCollection)
-    .removeProperty("key", (_path, property) => {
-      const idProp = getObjectProperty(j, _path.node, "id");
-      if (idProp === undefined && property.type === "ObjectProperty") {
-        const newIdProp = j.objectProperty(j.identifier("id"), property.value);
-        _path.node.properties.push(newIdProp);
-      }
-    })
-    .removeProperty("control")
-    .removeProperty("isFreeform")
-    .removeProperty("isToolSettings")
-    .removeProperty("isStatusBar")
-    .removeProperty("fillZone")
-    .removeProperty("syncEventIds")
-    .removeProperty("stateFunc")
-    .renameProperty("iconSpec", "icon")
-    .renameProperty("isFloatingStateSupported", "canFloat")
-    .renameProperty("badgeType", "badge")
-    .renameProperty("element", "content");
-}
-
-export function handleCanFloatProperty(j: JSCodeshift, objectToProperties: Map<ASTPath<ObjectExpression>, ObjectProperty[]>, canFloatProperty: string): Parameters<ObjectExpressionCollection["removeProperty"]>[1] {
-  return (path, property) => {
-    let objectProperties = objectToProperties.get(path);
-    if (!objectProperties) {
-      objectProperties = [];
-      objectToProperties.set(path, objectProperties);
-    }
-
-    if (property.type === "ObjectProperty") {
-      const newProperty = j.objectProperty(j.identifier(canFloatProperty), property.value);
-      objectProperties.push(newProperty);
-    }
-  };
 }
 
 export function takeFirstArrayExpressionElement(j: JSCodeshift, expression: ArrayExpression): ExpressionKind | undefined {
