+++ tools/codemod/src/v4.0.0/__testfixtures__/element-to-config/Frontstage.output.tsx
@@ -50,4 +50,4 @@
     size: 180,
     ...this.props.bottomPanelProps,
   },
-})
+});
