+++ tools/codemod/src/v4.0.0/__testfixtures__/element-to-config/FrontstageStagePanel.input.tsx
@@ -2,6 +2,7 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+import { Frontstage, StagePanel, Zone, Widget } from "@itwin/appui-react";
 <Frontstage
   id={this.id}
   version={1}
