+++ tools/codemod/src/v4.0.0/__testfixtures__/element-to-config/FrontstageStagePanel.output.tsx
@@ -88,4 +88,4 @@
       }],
     },
   },
-})
+});
